Implement reading messages and room topic events

+++ sync.go
@@ -80,7 +80,7 @@ type RoomEvent struct {
 	Type           string          `json:"type"`             // Required. The type of event. This SHOULD be namespaced similar to Java package naming conventions e.g. 'com.example.subdomain.event.type'
 	EventID        string          `json:"event_id"`         // Required. The globally unique event identifier.
 	Sender         string          `json:"sender"`           // Required. Contains the fully-qualified ID of the user who sent this event.
-	OriginServerTs int             `json:"origin_server_ts"` // Required. Timestamp in milliseconds on originating homeserver when this event was sent.
+	OriginServerTs int64           `json:"origin_server_ts"` // Required. Timestamp in milliseconds on originating homeserver when this event was sent.
 	Unsigned       UnsignedData    `json:"unsigned"`         // Contains optional extra information about the event.
 }
